agent/requirements.txt
livekit
livekit-agents>=0.12.9,<1
livekit-plugins-google>=0.10.0,<1
python-dotenv
fastapi
uvicorn
supabase
pydantic
asyncpg
httpx
